refactor(routes): update routes
