+++ front-end/src/index.js
@@ -9,19 +9,17 @@ import App from './App';
 import Login from './pages/login';
 import InitisaliseChat from './pages/Customer/InitialiseChat';
 import CustomerChat from './pages/Customer/CustomerChat';
-import FAQ from './pages/faq';
-import FaqIndivPage from './pages/FaqIndivPage';
-import AppointmentBooking from './pages/AppointmentBooking';
-import DetailedAppointmentBooking from './pages/DetailedAppointmentBooking';
-import StaffLandingPage from './pages/StaffLandingPage';
-import ViewBranchAppointments from './pages/ViewBranchAppointments';
-import ViewDetailedAppointments from './pages/ViewDetailedAppointments';
-import ViewOutstandingChatCustomers from './pages/ViewOutstandingChatCustomers';
-import AdminSupportLog from './pages/AdminSupportLog';
+import FAQSection from './pages/FAQSection';
+import FAQDetail from './pages/FAQDetail';
 
 // Staff Routes
-import StaffHome from './pages/Staff/StaffHome';
+import StaffLandingPage from './pages/Staff/StaffLandingPage';
 import StaffChats from './pages/Staff/StaffChats';
+import ViewBranchAppointments from './pages/Staff/ViewBranchAppointments';
+
+// Customer Appointment Booking Routes
+import AppointmentBooking from './pages/Customer/AppointmentBooking';
+import DetailedAppointmentBooking from './pages/Customer/DetailedAppointmentBooking';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
@@ -29,25 +27,26 @@ root.render(
     <Routes>
       <Route exact path="/" element={<App />} />
       <Route path="/login" element={<Login />} />
-      <Route path="/faq" element={<FAQ />} />
-      <Route path="/faq-article" element={<FaqIndivPage />} />
-      <Route path="/appointment-booking" element={<AppointmentBooking />} />
-      <Route path="/detailed-appointment-booking" element={<DetailedAppointmentBooking />} />
-      <Route path="/staff-landing" element={<StaffLandingPage />} />
-      <Route path="/login" element={<Login />} />
+
+      {/* FAQ Routes */}
+      <Route path='/faq/:section' element={<FAQSection />} />
+      <Route path='/faq/:section/:article' element={<FAQDetail />} />
+
+      {/* Chat Routes */}
       <Route path="/awaitchat" element={<InitisaliseChat />} />
       <Route path="/chat" element={<CustomerChat />} />
-      <Route path="/view-branch-appointments" element={<ViewBranchAppointments />} />
-      <Route path="/view-detailed-appointments" element={<ViewDetailedAppointments />} />
-      <Route path="/view-outstanding-chat-customers" element={<ViewOutstandingChatCustomers />} />
-      <Route path="/admin-support-log" element={<AdminSupportLog />} />
-
 
       <Route path="*" element={<h1>404 Not Found</h1>} />
 
-      {/* Protected Routes */}
-      <Route path="/staff/home" element={<ProtectedRoute Component={StaffHome} role="staff" />} />
+
+      {/* Protected Routes (Staff) */}
+      <Route path="/staff/home" element={<ProtectedRoute Component={StaffLandingPage} role="staff" />} />
       <Route path="/staff/chats" element={<ProtectedRoute Component={StaffChats} role="staff" />} />
+      <Route path="/staff/branches" element={<ProtectedRoute Component={ViewBranchAppointments} role="staff" />} />
+
+      {/* Protected Routes (Customer) */}
+      <Route path="/appointments/branches" element={<ProtectedRoute Component={AppointmentBooking} role="customer" />} />
+      <Route path="/appointments/timeslots" element={<ProtectedRoute Component={DetailedAppointmentBooking} role="customer" />} />
     </Routes>
   </BrowserRouter>
 );
